Style overhaul.

+++ steam-api-front/src/components/UserProfile.js
@@ -1,6 +1,4 @@
 import React, { Component } from 'react';
-import SelectField from 'material-ui/SelectField';
-import MenuItem from 'material-ui/MenuItem';
 import axios from 'axios';
 
 //The public page for a user, able to leave comments here as well as delete them if you're the owner of the comment.
@@ -24,7 +22,14 @@ class UserProfile extends Component {
     }
 
     handleClick(tar) {
-        axios.delete('http://localhost:8888/deleteComment/' + tar + ',' + localStorage.username)
+        let confirm = window.confirm('You sure you want to delete this comment?');
+        if (confirm) {
+            axios.delete('http://localhost:8888/deleteComment/' + tar + ',' + localStorage.username)
+            window.location = "http://localhost:3000/user/"
+        }
+        else {
+            console.log('Nice try guy!');
+        }
     }
 
     textChange(e) {
@@ -60,6 +65,7 @@ class UserProfile extends Component {
                 profileInfo: res.data.profileInfo,
                 loading: false
             })
+            console.log(this.state.profileInfo);
         })
     }
 
@@ -71,6 +77,26 @@ class UserProfile extends Component {
 
     render() {
         let profileInfo = this.state.profileInfo;
+        let popUp = (
+            <div>
+                <a class="waves-effect waves-light btn" href="#modal1">Modal</a>
+
+                <div id="modal1" class="modal">
+                    <div class="modal-content">
+                        <h4>Modal Header</h4>
+                        <p>A bunch of text</p>
+                    </div>
+                    <div class="modal-footer">
+                        <a href="#!" class=" modal-action modal-close waves-effect waves-green btn-flat">Agree</a>
+                    </div>
+                </div>
+            </div>)
+        let style = {
+            backgroundColor: '#FFFFFF',
+            backgroundImage: 'url(\'https://images4.alphacoders.com/681/681185.jpg\')',
+            backgroundPositionY: '-10px',
+            backgroundRepeat: 'no-repeat'
+        }
 
         if (this.state.loading) {
             return <div className="progress">
@@ -79,37 +105,43 @@ class UserProfile extends Component {
         }
         else {
             return (
-                <div className="container">
-                    <h2>Welcome to {localStorage.userSearch}'s Page!</h2>
 
-                    <div>
-                        <h5>About</h5>
-                        <p>Age {this.state.age}</p>
-                        <p>Member since {this.state.accountCreated.replace(/T.*?Z/, '')}</p>
-                        <p>{this.state.userBio}</p>
+                <div className="JSXWrapper">
+                    <div style={style} className="jumbotron">
+                        <h2>WELCOME TO {localStorage.userSearch}'s PAGE!</h2>
                     </div>
+                    <div className="container">
 
-                    <div>
-                        <p>Leave a comment!</p>
-                        <form onSubmit={this.handleSubmit}>
-                            <input className="gameInput" placeholder="Wanna join up? Leave them your details!" type="text" onChange={this.textChange} onSubmit={this.handleSubmit} />
-                        </form>
-                    </div>
-
-
-                    {profileInfo.map((profile) =>
                         <div>
-                            <ul className="collection">
-                                <li className="collection-item avatar">
-                                    <span className="title">From: {profile.sender_id}</span>
-                                    <p className="truncate">{profile.comment}</p>
-                                    <button onClick={() => { this.handleClick(profile.id) } } >Delete</button>
-                                </li>
-                            </ul>
+                            <h5>ABOUT</h5>
+                            <p>AGE: {this.state.age}</p>
+                            <p>MEMBER SINCE: {this.state.accountCreated.replace(/T.*?Z/, '')}</p>
+                            <p>{this.state.userBio}</p>
                         </div>
-                    )}
+
+                        <div>
+                            <p>Leave a comment!</p>
+                            <form onSubmit={this.handleSubmit}>
+                                <input className="gameInput" placeholder="Wanna join up? Leave them your details! (Eg. Steam ID, battle.net tag)" type="text" onChange={this.textChange} onSubmit={this.handleSubmit} />
+                            </form>
+                        </div>
+
+
+                        {profileInfo.map((profile) =>
+                            <div>
+                                <ul className="collection">
+                                    <li className="collection-item avatar">
+                                        <img src="http://www.conquerthewallmarathon.com/wp-content/uploads/2014/05/generic-user-icon.jpg" className="circle" />
+                                        <span className="title">From: {profile.sender_id}</span>
+                                        <p className="truncate">{profile.comment}</p>
+                                        {this.state.onPage === profile.sender_id ? <button className="btn danger waves-effect waves-light right-align" onClick={() => { this.handleClick(profile.id) } } ><i className="material-icons">delete</i>Delete</button> : null}
+                                    </li>
+                                </ul>
+                            </div>
+                        )}
 
                 </div>
+            </div>
             )
         }
     }
